Permissions added

+++ token_authentication/views.py
@@ -1,22 +1,24 @@
+from .permissions import UsersObservationPermission
 from rest_framework import status
 from rest_framework.generics import RetrieveUpdateDestroyAPIView
-from rest_framework.permissions import AllowAny, IsAuthenticated
+from rest_framework.permissions import AllowAny, IsAuthenticated, IsAdminUser
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from rest_framework.viewsets import ModelViewSet
 
 from .renders import UserJSONRenderer
-from .serializers import (
-    LoginSerializer, RegistrationSerializer, UserSerializer,
-)
+from .serializers import LoginSerializer, RegistrationSerializer, UserSerializer, AdminViewAllSerializer
+from .models import User
 
 
 class RegistrationAPIView(APIView):
-    permission_classes = (AllowAny,)
+    permission_classes = (IsAdminUser,)
     serializer_class = RegistrationSerializer
     renderer_classes = (UserJSONRenderer,)
 
     def post(self, request):
         user = request.data.get('user', {})
+
         serializer = self.serializer_class(data=user)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -39,24 +41,34 @@ class LoginAPIView(APIView):
 
 
 class UserRetrieveUpdateDestroyAPIView(RetrieveUpdateDestroyAPIView):
-    permission_classes = (IsAuthenticated,)
+    permission_classes = (IsAuthenticated, UsersObservationPermission,)
     renderer_classes = (UserJSONRenderer,)
     serializer_class = UserSerializer
 
     def retrieve(self, request, *args, **kwargs):
-        serializer = self.serializer_class(request.user)
+        serializer = self.serializer_class(User.objects.get(pk=kwargs['pk']))
+        self.check_object_permissions(request, serializer.instance)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def update(self, request, *args, **kwargs):
         serializer_data = request.data.get('user', {})
-        serializer = self.serializer_class(request.user, data=serializer_data, partial=True)
+        serializer = self.serializer_class(User.objects.get(pk=kwargs['pk']), data=serializer_data, partial=True)
+        self.check_object_permissions(request, serializer.instance)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def destroy(self, request, *args, **kwargs):
-        serializer = self.serializer_class(request.user)
+        serializer = self.serializer_class(User.objects.get(pk=kwargs['pk']))
+        self.check_object_permissions(request, serializer.instance)
         instance = serializer.instance
         instance.delete()
         return Response(serializer.data, status=status.HTTP_200_OK)
 
+
+class AllUsersView(ModelViewSet):
+    queryset = User.objects.all()
+    permission_classes = (IsAuthenticated,)
+
+    def get_serializer_class(self):
+        return AdminViewAllSerializer if self.request.user.is_staff else UserSerializer
